refactor: Refactor imgbuf logic

+++ src/draw/draw_pixel.c
@@ -6,5 +6,5 @@ void	draw_pixel(t_img_info *img_info, int x, int y, int color)
 
 	color_byte = img_info->bpp / 8;
 	pixel = (y * img_info->size_line) / color_byte + x;
-	*(img_info->buf + pixel) = (unsigned int)color;
+	*(img_info->buf + pixel) = color;
 }
